refactor: detach util functions
