+++ utils/getCurrentTimestamp.ts
@@ -1,0 +1,3 @@
+export default function getCurrentTimestamp(): number {
+  return ~~(Date.now() / 1000);
+};
