+++ utils/getSumOfNumberArray.ts
@@ -1,0 +1,3 @@
+export default function getSumOfNumberArray(arrayOfNumbers: number[]): number {
+  return arrayOfNumbers.reduce((a, b) => a + b, 0);
+}
