+++ utils/removeLinebreaksFromText.ts
@@ -1,0 +1,6 @@
+export default function removeLinebreaksFromText(text: string): string {
+  return text
+    .split('\n')
+    .filter(isNotEmpty => isNotEmpty)
+    .join(' ');
+};
